Move server testing infrastructure to test sources
