+++ src/test/java/org/springframework/http/server/reactive/bootstrap/ReactorHttpServer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.http.server.reactive.boot;
+package org.springframework.http.server.reactive.bootstrap;
 
 import reactor.core.flow.Loopback;
 import reactor.core.state.Completable;
