+++ src/test/java/org/springframework/http/server/reactive/bootstrap/TomcatHttpServer.java
@@ -14,25 +14,37 @@
  * limitations under the License.
  */
 
-package org.springframework.http.server.reactive.boot;
+package org.springframework.http.server.reactive.bootstrap;
 
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.ServerConnector;
-import org.eclipse.jetty.servlet.ServletContextHandler;
-import org.eclipse.jetty.servlet.ServletHolder;
+import java.io.File;
+
+import org.apache.catalina.Context;
+import org.apache.catalina.LifecycleException;
+import org.apache.catalina.startup.Tomcat;
 
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.http.server.reactive.ServletHttpHandlerAdapter;
 import org.springframework.util.Assert;
 
+
 /**
  * @author Rossen Stoyanchev
  */
-public class JettyHttpServer extends HttpServerSupport implements InitializingBean, HttpServer {
+public class TomcatHttpServer extends HttpServerSupport implements InitializingBean, HttpServer {
 
-	private Server jettyServer;
+	private Tomcat tomcatServer;
 
 	private boolean running;
+
+	private String baseDir;
+
+
+	public TomcatHttpServer() {
+	}
+
+	public TomcatHttpServer(String baseDir) {
+		this.baseDir = baseDir;
+	}
 
 
 	@Override
@@ -43,30 +55,32 @@ public class JettyHttpServer extends HttpServerSupport implements InitializingBe
 	@Override
 	public void afterPropertiesSet() throws Exception {
 
-		this.jettyServer = new Server();
+		this.tomcatServer = new Tomcat();
+		if (this.baseDir != null) {
+			this.tomcatServer.setBaseDir(baseDir);
+		}
+		this.tomcatServer.setHostname(getHost());
+		this.tomcatServer.setPort(getPort());
 
 		Assert.notNull(getHttpHandler());
 		ServletHttpHandlerAdapter servlet = new ServletHttpHandlerAdapter();
 		servlet.setHandler(getHttpHandler());
-		ServletHolder servletHolder = new ServletHolder(servlet);
 
-		ServletContextHandler contextHandler = new ServletContextHandler(this.jettyServer, "", false, false);
-		contextHandler.addServlet(servletHolder, "/");
-
-		ServerConnector connector = new ServerConnector(this.jettyServer);
-		connector.setHost(getHost());
-		connector.setPort(getPort());
-		this.jettyServer.addConnector(connector);
+		File base = new File(System.getProperty("java.io.tmpdir"));
+		Context rootContext = tomcatServer.addContext("", base.getAbsolutePath());
+		Tomcat.addServlet(rootContext, "httpHandlerServlet", servlet);
+		rootContext.addServletMapping("/", "httpHandlerServlet");
 	}
+
 
 	@Override
 	public void start() {
 		if (!this.running) {
 			try {
 				this.running = true;
-				this.jettyServer.start();
+				this.tomcatServer.start();
 			}
-			catch (Exception ex) {
+			catch (LifecycleException ex) {
 				throw new IllegalStateException(ex);
 			}
 		}
@@ -77,10 +91,10 @@ public class JettyHttpServer extends HttpServerSupport implements InitializingBe
 		if (this.running) {
 			try {
 				this.running = false;
-				jettyServer.stop();
-				jettyServer.destroy();
+				this.tomcatServer.stop();
+				this.tomcatServer.destroy();
 			}
-			catch (Exception ex) {
+			catch (LifecycleException ex) {
 				throw new IllegalStateException(ex);
 			}
 		}
